+++ test/function_test.py
@@ -10,7 +10,7 @@ from synchronicity.exceptions import UserCodeException
 import modal
 from modal import Image, Mount, NetworkFileSystem, Proxy, Stub, web_endpoint
 from modal._vendor import cloudpickle
-from modal.exception import DeprecationError, ExecutionError, InvalidError
+from modal.exception import ExecutionError, InvalidError
 from modal.functions import Function, FunctionCall, gather
 from modal.runner import deploy_stub
 from modal_proto import api_pb2
@@ -35,34 +35,17 @@ def dummy():
 def test_run_function(client, servicer):
     assert len(servicer.cleared_function_calls) == 0
     with stub.run(client=client):
-        # Old-style remote calls
-        with pytest.raises(DeprecationError):
-            foo.call(2, 4)
-
-        # New-style remote calls
         assert foo.remote(2, 4) == 20
         assert len(servicer.cleared_function_calls) == 1
 
 
 @pytest.mark.asyncio
 async def test_call_function_locally(client, servicer):
-    # Old-style local calls
-    with pytest.raises(DeprecationError):
-        foo(22, 44)
-
-    with pytest.raises(DeprecationError):
-        await async_foo(22, 44)
-
-    # New-style local calls
     assert foo.local(22, 44) == 77  # call it locally
     assert await async_foo.local(22, 44) == 78
 
     with stub.run(client=client):
         assert foo.remote(2, 4) == 20
-        with pytest.raises(DeprecationError):
-            foo(22, 55)
-        with pytest.raises(DeprecationError):
-            await async_foo(22, 44)
         assert async_foo.remote(2, 4) == 20
         assert await async_foo.remote.aio(2, 4) == 20
 
@@ -216,10 +199,6 @@ async def test_generator(client, servicer):
         assert next(res) == "bar"
         assert list(res) == ["baz", "boo"]
         assert len(servicer.cleared_function_calls) == 1
-
-        # Check deprecated interface
-        with pytest.raises(DeprecationError):
-            later_gen_modal.call()
 
 
 @pytest.mark.asyncio
@@ -490,13 +469,6 @@ def test_allow_cross_region_volumes_webhook(client, servicer):
                 assert svm.allow_cross_region
 
 
-def test_shared_volumes(client, servicer):
-    stub = Stub()
-    vol = NetworkFileSystem.from_name("foo")
-    with pytest.raises(DeprecationError):
-        stub.function(shared_volumes={"/sv-1": vol})
-
-
 def test_serialize_deserialize_function_handle(servicer, client):
     from modal._serialization import deserialize, serialize
 
@@ -525,7 +497,8 @@ def test_default_cloud_provider(client, servicer, monkeypatch):
     monkeypatch.setenv("MODAL_DEFAULT_CLOUD", "oci")
     stub.function()(dummy)
     with stub.run(client=client):
-        f = servicer.app_functions[stub.dummy.object_id]
+        object_id: str = stub.indexed_objects["dummy"].object_id
+        f = servicer.app_functions[object_id]
 
     assert f.cloud_provider == api_pb2.CLOUD_PROVIDER_OCI
 
@@ -574,7 +547,8 @@ def test_deps_explicit(client, servicer):
     stub.function(image=image, network_file_systems={"/nfs_1": nfs_1, "/nfs_2": nfs_2})(dummy)
 
     with stub.run(client=client):
-        f = servicer.app_functions[stub.dummy.object_id]
+        object_id: str = stub.indexed_objects["dummy"].object_id
+        f = servicer.app_functions[object_id]
 
     dep_object_ids = set(d.object_id for d in f.object_dependencies)
     assert dep_object_ids == set([image.object_id, nfs_1.object_id, nfs_2.object_id])
@@ -605,22 +579,22 @@ def assert_is_wrapped_dict(some_arg):
 
 
 def test_calls_should_not_unwrap_modal_objects(servicer, client):
-    stub = Stub()
-    stub.some_modal_object = modal.Dict.from_name("blah", create_if_missing=True)
+    some_modal_object = modal.Dict.lookup("blah", create_if_missing=True, client=client)
 
+    stub = Stub()
     foo = stub.function()(assert_is_wrapped_dict)
     servicer.function_body(assert_is_wrapped_dict)
 
     # make sure the serialized object is an actual Dict and not a _Dict in all user code contexts
     with stub.run(client=client):
-        assert type(foo.remote(stub.some_modal_object)) == modal.Dict
-        fc = foo.spawn(stub.some_modal_object)
+        assert type(foo.remote(some_modal_object)) == modal.Dict
+        fc = foo.spawn(some_modal_object)
         assert type(fc.get()) == modal.Dict
-        for ret in foo.map([stub.some_modal_object]):
+        for ret in foo.map([some_modal_object]):
             assert type(ret) == modal.Dict
-        for ret in foo.starmap([[stub.some_modal_object]]):
+        for ret in foo.starmap([[some_modal_object]]):
             assert type(ret) == modal.Dict
-        foo.for_each([stub.some_modal_object])
+        foo.for_each([some_modal_object])
 
     assert len(servicer.client_calls) == 5
 
@@ -631,16 +605,16 @@ def assert_is_wrapped_dict_gen(some_arg):
 
 
 def test_calls_should_not_unwrap_modal_objects_gen(servicer, client):
-    stub = Stub()
-    stub.some_modal_object = modal.Dict.from_name("blah", create_if_missing=True)
+    some_modal_object = modal.Dict.lookup("blah", create_if_missing=True, client=client)
 
+    stub = Stub()
     foo = stub.function()(assert_is_wrapped_dict_gen)
     servicer.function_body(assert_is_wrapped_dict_gen)
 
     # make sure the serialized object is an actual Dict and not a _Dict in all user code contexts
     with stub.run(client=client):
-        assert type(next(foo.remote_gen(stub.some_modal_object))) == modal.Dict
-        foo.spawn(stub.some_modal_object)  # spawn on generator returns None, but starts the generator
+        assert type(next(foo.remote_gen(some_modal_object))) == modal.Dict
+        foo.spawn(some_modal_object)  # spawn on generator returns None, but starts the generator
 
     assert len(servicer.client_calls) == 2