+++ test/object_test.py
@@ -10,27 +10,9 @@ async def test_async_factory(client):
     stub = Stub()
     with pytest.warns(DeprecationError):
         stub.my_factory = Queue.new()
-    async with stub.run(client=client):
-        assert isinstance(stub.my_factory, Queue)
-        assert stub.my_factory.object_id == "qu-1"
-        with pytest.raises(DeprecationError):
-            stub.app.my_factory
-
-
-@pytest.mark.asyncio
-async def test_use_object(client):
-    # Deploy object
-    q = await Queue.lookup.aio("foo-queue", create_if_missing=True, client=client)
-
-    # Use object
-    stub = Stub()
-    q = Queue.from_name("foo-queue")
-    assert isinstance(q, Queue)
-    stub.my_q = q
-    async with stub.run(client=client):
-        assert stub.my_q.object_id == "qu-1"
-        with pytest.raises(DeprecationError):
-            stub.app.my_q
+        async with stub.run(client=client):
+            assert isinstance(stub.my_factory, Queue)
+            assert stub.my_factory.object_id == "qu-1"
 
 
 def test_new_hydrated(client):