+++ test/runner_test.py
@@ -20,11 +20,17 @@ def test_run_stub(servicer, client):
     ctx.pop_request("AppClientDisconnect")
 
 
+def dummy():
+    ...
+
+
 def test_run_stub_profile_env_with_refs(servicer, client, monkeypatch):
     monkeypatch.setenv("MODAL_ENVIRONMENT", "profile_env")
     with servicer.intercept() as ctx:
         dummy_stub = modal.Stub()
-        dummy_stub.ref = modal.Secret.from_name("some_secret")
+        ref = modal.Secret.from_name("some_secret")
+        dummy_stub.function(secrets=[ref])(dummy)
+
     assert ctx.calls == []  # all calls should be deferred
 
     with servicer.intercept() as ctx:
@@ -45,10 +51,12 @@ def test_run_stub_profile_env_with_refs(servicer, client, monkeypatch):
 def test_run_stub_custom_env_with_refs(servicer, client, monkeypatch):
     monkeypatch.setenv("MODAL_ENVIRONMENT", "profile_env")
     dummy_stub = modal.Stub()
-    dummy_stub.own_env_secret = modal.Secret.from_name("own_env_secret")
-    dummy_stub.other_env_secret = modal.Secret.from_name(
+    own_env_secret = modal.Secret.from_name("own_env_secret")
+    other_env_secret = modal.Secret.from_name(
         "other_env_secret", environment_name="third"
     )  # explicit lookup
+
+    dummy_stub.function(secrets=[own_env_secret, other_env_secret])(dummy)
 
     with servicer.intercept() as ctx:
         ctx.add_response("SecretGetOrCreate", api_pb2.SecretGetOrCreateResponse(secret_id="st-123"))